fix(node-core,blockchain-tree): port to v2.4.1 API - fix imports, missing deps, and BlockBuffer
reth-node-core:
- Remove reth-engine-local dep (breaks node-api cycle)
- Add missing deps: reth-consensus, reth-primitives-traits, reth-rpc-convert,
  reth-net-banlist, reth-tasks, reth-engine-primitives, reth-tracing-otlp, url, ipnet
- Replace payload_builder.rs with v2.4.1 version (DefaultPayloadBuilderValues,
  gas_limit/max_blobs optional, no ETHEREUM_BLOCK_GAS_LIMIT constant)
- Add DefaultPruningValues + PruneConfigKind to pruning.rs
- Add StaticFilesConfig/BlocksPerFileConfig to reth-config
- Add DEFAULT_CROSS_BLOCK_CACHE_SIZE_MB + re-export DEFAULT_* from engine-primitives
  in node_config.rs
- Fix Head struct construction (use BlockHeader trait methods, drop header_td_by_number)
- Fix fetch_tip_from_network to be generic over Client::Header
- Fix Cow<str> -> OsStr conversions in rpc_server.rs/txpool.rs
- Fix cross_block_cache_size u64->usize cast in engine.rs
- Fix with_dev const fn (DevArgs has a destructor)

reth-blockchain-tree:
- Replace lib.rs with minimal stub (re-export reth_blockchain_tree_api::*, BlockBuffer)
- Add block_buffer.rs with BlockBuffer copied from engine-tree (avoids circular dep
  since engine-tree depends on blockchain-tree, not vice versa)
- Add deps: reth-primitives-traits, alloy-consensus, indexmap

+++ crates/node/core/src/args/engine.rs
@@ -263,7 +263,7 @@ impl Default for DefaultEngineValues {
             state_cache_disabled: false,
             prewarming_disabled: false,
             state_provider_metrics: false,
-            cross_block_cache_size: DEFAULT_CROSS_BLOCK_CACHE_SIZE_MB,
+            cross_block_cache_size: DEFAULT_CROSS_BLOCK_CACHE_SIZE_MB as usize,
             state_root_task_compare_updates: false,
             accept_execution_requests_hash: false,
             multiproof_chunk_size: DEFAULT_MULTIPROOF_TASK_CHUNK_SIZE,
